First sqlite migration.
In this migration we introduce support for sqlite database
backing.

We needed to change quite some migration files so that they only
conditionally apply to mssql, as sqlite3 does not support some features.
To check which ones which ones [1].

Also we needed to finally migrate the Logs database through the ef
and we broke the logging by Serilog. I do not know why nor can i see
any debugging but this is secondary for now as long as the console
logging is working.

The dotnet docker file was also modified so that sqlite3 is installed.

[1] https://docs.microsoft.com/en-us/ef/core/providers/sqlite/limitations