Login.cshtml: Redirect to account on Login.
We cannot redirect to Map on Login when we have a virgin database,
where there are no devices. In such case the DeviceController will
not provide routes because there are no devices. This became an
exception, correctly, with core 2.2. Now we redirect to the account
profile, which is guaranteed to exist if there was a successful login.