docker-compose: Do not take responsibility for database containers.
mssql is now optional and it is not easy to have migrations happening
on a mssql server from scratch. For the complex deployments where mssql
is required it is assumed that the database is an independent block.

On the other hand by default we prefer sqlite, which allows for a completely
self-contained deployment with the migrations included.