applications.json: As mentioned before mssql logging is broken. Disable it in configuration.
Since we have started using the Logs as an entity model serilog logging
to the Logs database is broken. Furthermore we were asking serilog to create
the Logs table in the mssql configuration which led to a long hanging startup
with no information when no mssql server was available. As this is currently
the default this led to very long startup time in production configurations.