docker-dotnet: Apply migrations with sqlite3 database provider.
We can only do build time migration for the case of sqlite database
provider, as the sqlite3 database is packaged with the dotnet container.
We force it so that even if in the future the default database provider
changes, we will not have docker builds failing, trying to migrate into
a mssql server which probably does not exist.