Map.cshtml: Add a handler for DroneReply telemetry.
This is for now a bit stupid as there is no way for the frontend
to know what the DroneReply refers to. This fake handler can in
the future added to handle ping services, but for now this is not
the case. The reason we need to add this handler is because we are
currently having a common code for Drone Service that records and
forwards all the drone service's messages.