parrot-service: Add and use LogHelper to include class name
Instead of relying on a global logging facility we now have a LogHelper
class that when a class inherits, provides a logger that sets the name
to the class name. Together with the basicConfig change this now allows
for the log to contain the class to which the log message relates to.