parrot-service: Make TWebSocketsClient a managed resource.
We need to make TWebSocketsClient a managed resource because we are not letting
Thrift control the open and close of the transport. The reason we do this is that
the pooled thead servers open and close the client transport with each accept.
This is unnaceptable with our concept as this would mean no connection reuse."