parrot-service: Improved process return code handling.
The getVideoState method of DroneVideoContainer was not handling
all the cases of failure. For example the test for != None was
true when the processes were terminated by POSIX signals and
the method then returned the process was alive when it actually
was not.
Also inverted the if structure to allow for polling only once
per if-else clause instead of every condition possible.