DroneMicroServiceManager: Corrected GatherVideoSource logic.
Previously we were not checking the result of the sendVideoToAsync at
all which led to a single request for the video and then just recording
whether the video was dead or not. This situation was pretty useless as
there was no way to recover the video if a drone stopped sending video for
any reason.

Also consolidated the logic into one, where we always try to get the video
state before we try to ask for any video. This also adds some sanity and
simplifies a lot the logic.