Added support for configurable GoogleMapsAPI key.
WE had previously hardcoded the (which is a bad practice), but this
had the problem that the key is associated with a domain, which is
not localhost or non production addresses. As we normally do not need
the key for development the need for a configurable key rose. To this
end the GoogleMapsAPIKey setting was added to the RepositoriesConfiguration.
The Map.csthml and the DeviceRepositoryController now use this settings
to control the key that gets rendered in the user.

A model class was added called RepositoryOptions...i know this seems a duplication
but this is the way to go, and the other RepositorySettings and the DropboxSettings
will soon be merged into the RepositoryOptions model.