Drohub: Created a many2many relation between Users and Devices.
The logic is that many users may have control over a device,
but may for example have different roles and permissions. This
does not exist now, but the many 2 many relationship must exist
to allow these kinds of scenarios.

Several changes happened, namely that a device may already exist
but not visible for the user. While before when the user created
a device, it was only visible for him, now this is not the case.

This led to different semantics where for example the user may
delete his association with a device, but that does not mean
that the device is deleted, as it may still be in use by others.

Also, entity frameworks does not allow for a direct many to many
relationship, and instead we needed to create an intermediate table
that maps User and Devices. This turned out to be quite a better
way of visualizing the relation.

Very good information on [1]

[1] https://www.thereformedprogrammer.net/updating-many-to-many-relationships-in-entity-framework-core/