database: Replaced sqlite3 with mysql.
We need to go this way due to the extremely limited migrations
support of sqlite3. This could cost us very dearly in the future
and with mysql being free there is not much reason to not go this
way. It is true that sqlite3 is much lighter in resource usage but
we cannot afford the engineering risk.