Program: Added support for auto migration.
As we moved away from the sqlite3 for good we need to improve
our deployment strategy. With this in mind we have added an
auto migrate ability turned on by default for the production
configuration. THis is required because it is possible that
a fresh docker deployment needs to update the independent database.