docker-compose: Added support for mysql service.
The mysql service is now our default database provider and
as such we need to orchestrate the database service start
alongside the drohub website. Lots of default options are set
including security ones. Future work should make these security
options overridable.

Also due to the fact that the database might not be up by the
time the DroHub web server starts we need set up a policy of restart
on failure. This way it will retry to start until the connection is
up. This is also useful for crashes of the webserver.