DatabaseMigrationExtension: IWebHostExtensions is now partial.
We do not want to create a class for every extension as they
all have the same general object of work. So we make the
IWebHostExtensions a partial class where other methods may be
added separately. This seems a lot like using the class as a
namespace and in a certain sense it is...C# and it's million
syntactic sugars or my ignorance :)