DatabaseMigrationExtension: Fail on migration failure.
After some thought I think I would rather the server did not try
to start with a migration done half way. It will probably crash
the server but that is actually a good thing, as if we do not
have a valid database to make the migrations the server is useless
regardless.