InitializeAdminUserExtension: Create an admin user on startup.
As we disabled the creation of users it is important that there
is always a system user. For this case the system user is called
admin and it's password is created automatically and shown in the
logs for the admin user to see and modify it as soon as possible.

There was a small adjustment on Program.cs file so that we are
able to factor in the need for this new IWebHost extension.