InitializeAdminUserExtension: Added IsAdminPolicy and Role:Admin claim.
In this commit we add an authorization policy that allows the
checking of whether a given user is an administrator or not.
To allow this we also needed to create a Role:Admin claim, that
is set by default if the user name is the AdminUserName.

Some re-organization of the code was done to have the AddIsAdminPolicy
local to the InitializeAdminUserExtensions file and still visible in the
IdentityHostingStartup class.