User creation: Allow user creation from the AdminPanel.
Moved the logic of user creation out of the main page into a special
page in the account management. This should later be made restricted
to special roles or policies.

The class names were not changed yet. This will be done when the work
on fleshing the AdminPanel will start.