modals: Added ability to render modals from partial views.
This ability is crucial to have a simpler website navigation
while keeping the organization facilities that dotnet provides
us.

An important detail of how this works is that when the modelstate
fails the modal does not exit, instead showing the model state errors.
All this with minimal javascript. Very happy :)