Device: Create device is now a partial view on a modal.
In this commit we make the Create device view a partial view that
is to be used as a modal view. This makes the device creation simpler
and does not make the user need to leave the current page. There is
a small bug where after the device is created, the sidebar list does
not update. Not very worried about this as this sidebar list will probably
be removed later on.