InitializeAdminUserExtension log a machine readable message.
We may need the admin password printed in the logs to be machine readable.
The way it was, it was not easy to read. Now we have a magic string GENERATE ROOT PASSWORD
and after it the actual password. The message then has a line ending so that the
loggers additional metadata does not make the message hard to parse...Not ideal still
but that is life.