DroHubTests: Added Integration Tests.
The integration tests in this stage spin up the docker-compose that
represents the infrastructure of DroHub and runs tests. Right now the
all the docker images are build with sources from this repository and
the tests happen with virgin containers. This may be annoying and unnecessary
for developers as it may take too long to run simple tests. To make it
better the developer can comment out the disposal of the containers or
the force build in the TestServerFixture class.