IntegrationTest: Refactored TestLogin and added login helpers.
Added login helpers in a class called HttpClientHelper that allow to
login either as a specific user or as an administrator. With this helper
it was easy to add more login validation tests.

As a side effect of the HttpClientHelper code refactor the test of the login
page also got simpler.