ThriftMessageHandler: Temporarily disable content type check
We need to disable it because we cannot test this. When we try to
set the Content Type in the Integration tests they are silently
ignored. This leads to tests not working. I opened a ticket for this [1]

Even so i left the content type set in the IntegrationTest, so
that when we enable the content type check again the tests work
automatically.

https://github.com/dotnet/runtime/issues/1035