IntegrationTest: Removed duplicate assumption in TestCreateAndDelete.
Tests in xUnit are ran out of order so it happens
that running the tests expecting a side effect from a previous
test will not work. To then have repeatable tests we need to
delete the device we created for the test. For this we added
methods that allow us to delete the device after we created it.