TestThriftDrone: Finalized the changes required for this test.
This test mocks a drone service with the exception of the the
actual video as well as action verification. This will be
done in a later moment.

In this commit we needed to add an API to the DevicesController
so that we can retrieve telemetry list. This is useful for the
tests to validate that the simulated information actually got
written and is available, as well as providing an endpoint for
telemetry analysis by a frontend.

A signalR client is also used to verify that the mocked device
information is propagated into the signalR client frontends.

Some small removals of debugging information were also done
in components required for the drone mocking.