version hash. Version hash now correctly appears in the docker image again.
Due to a wrong assumption in the .dockerignore file the file
containing the hash was not being put in the docker build workspace.
This led to silent empty hash version being put in the image. To
avoid this happening again we added a test that checks that the hash
is there. To achieve this we also tweaked the footer markup so that
the hash is inside a span tag that we can track and possibly in the
future style.