Dashboard.cshtml: Refactored device action code and markup.
With the goal of separating the presentation markup from
the code required for the backend we are now moving towards
data-* tags which have functionality backed by the site.js
javascript.

For the device actions this means that no longer we have the
onclick attribute with inline javascript. Instead the logic
for followActionAndDoNothing is associate with the data-toggle
ajax-request.