dashboard: Complete javascript fronend re-factor
Removed all but one small js function from the Dashboard.cshtml
and moved it all to dashboard.cs. The object of the refactor
is such that we have closures instead of imperative function.

We also make the html pass data to the javascript logic through
data-\* attributes.

In the end we have 4 classes in this factoring: SignalRConnectionClass
responsible for setting up the connection to the SignalR hub.

JanusVideoClass responsible for setting up the connection between
the frontend and the Janus video backend.

TelemetryClass responsible for rendering and updating telemetry
from both initial sources as well as updates from the SignalR connection.

ModalClass responsible for handling the gap between the partial
views into modals.

All the dropbox javascript was removed as it is not currently a
priority and was giving a massive rot smell that would not be
fixed soon.