IdentityAuthorization: Added AuthorizationOptionsExtension.
Due to the fact that we cannot have access to the policies available
and registered in our system we need to have extensions that allow
for the recording of existing policies before passing them to the
policy subsystem.
This is a bit ugly and fragile but I removed the AddAuthorization from
the IdentityHostingStartup in an effort to make it obvious that
policies should not be registered directly with the Core  API
but through our 'configurator' extension.