AdminPanel: Added a basic user data table.
This is the first work on the dashboard that add meaningful administration
functionality. At this point we can see all the users registered in the system.
For each user we are able to check the policies they have associated, last login
time, devices and if they are locked out.

It is a bit ugly but this not the point at this point.