DroneMicroServiceManager: Added mountpoint cleanup to GatherVideoSource.
After the video mount point was created in Janus there was a moment
between the actual drone action and the mountpoint creation, where
an exception could escape without triggering the cleanup of the
created mountpoint. This would lead to a resource leak, as a dangling
mountpoint. Now we catch the exception, cleanup the mountpoint and
rethrow the exception.