backend: Added support for automatic live video recording.
In this commit we setup a system where whenever there is a video mountpoint
created due to a new device being connected, the received video starts
being recorded to a volume that is attached to both the janus and web
container. The paths of the docker-compose and the appsettings.json
are aligned so that the videos are shared between the containers.

The format of the recording is mjr, which is a janus format containing
a dump of the rtp frames sent by a drone device. No frontend exists
yet so these files while reachable through the backend server, are not
reachable to the user yet.