ThriftMessageHandler: Simplified websocket handling and blocking.
We no longer rely on a task completion source to block
but instead block and wait for the tasks that work on
the data from the websocket to complete. This simplifies
the code for handler finalization and treats the ReceiveFromWebSocket
as any other task.

ReceiveFromWebSocket was changed to reflect the new blocking
realities and hopefuly also improved the connection handing
specially on close.