Dashboard: Added support for video fullscreen.
The video overlay now has a fullscreen toggle button. The
video overlay is preserved when in fullscreen.

Also added some css classes to position elements in the
video overlay.

The device-video-overlay now sets sane defaults for overlay
elements. We moved the defaults from the container-action-overlay
to the device-video-overlay because then non-action elements
can also benefit from the defaults.