dashboard.js Unified class range telemetry rendering.
A lot of the logic that added or removed element classes
based on some information was coded with if clauses
all over the rendering code. Now we have a new class
called ElementClassRangeClass that allows us to add
element class attributes according to a value from a
range.

Refactored code to use this new code. As a bonus, also
got more powerful class add/remove control.