Added MJR-Converter docker service to support timeline video gallery.
This is a big commit which introduces the ability for the user
to see recordings of the live videos. This is useful for when
something happens to the drone and the only way to recover video
data is to get a replay of what you got through the livestream.

To accomplish this we need to have a way that converts the recordings
that janus produces, in mjr, and convert them to webm files. To
achieve this, this commit includes a docker service which listens
to mjr.tmp files being renamed to .mjr. When this happens it means
that the mjr files were closed and are final. The MJR-converter service
then post-processes the videos and makes them available as webm files.
The original mjr files are removed by default if the conversion was
successful.

All this files are then available through a timeline gallery page
where they can be viewed, and seeked if necessary.

The layout of the Gallery is very simple for now but we need input
from the designer before more work is applied.