parrot-service CopterHUDActivity: Added activity, and auxiliary assets.
This commit is basically a paste of all the assets required to
have the CopterHudActivity from parrot's GroundSdkDemo.

At this moment the thrift telemetry is not working but as
the project is in a state of flux at this point this does not
matter. When all the parts are in place the respective connections
will be enabled.