Integration of android video with drohub.
Glued the application logic so that we are able to broadcast
drone video from the android device to the Drohub platform.

This is the first alpha integration on which we will build upon.
In this version we integrate the video room concept of video signaling
to connect the android device with the drohub, removing the usage of the
video streaming concept. We broke the video recording in the process and
this needs to be put to work again.

Also working poorly is the THriftConnection libraries on the android
side. THey are unreliable at this time. This is an improvement even so
because the previous state was completely broken in regards to reliability.