parrot-service: Re-implemented TReverseTunnel and cleanups.
Syntax sugar cleanup in ThriftConnection class.
TReverseTunnelServer has been rewritten to use a
producer/consumer pattern. FOr now this is not very
canonical because we discard the fake accepted connection
with every operation end but this does not need to the case
and thrift actually expects to re-use the connection. Alas,
it also works this way, and the pattern is much safer.