DroneMicroServiceManager: Do not re-trigger the send video if it is STOPPED.
If the video state is stopped it means that the drone
backend is getting ready for video or does not want to
broadcast. An example of such a rejection is if the user
does not give screen cast permissions. We should honor that
and not spam the user until he joins.