live video drohub: Automatic starting of live video if available.
When our device says we are live we get a signalr
notification. Instead of relying on information from janus
we use our telemetry render concept and just render the live
video based on the telemetry. One important detail is that
if the render does not start in 10s we consider the render
stopped and try it again on the next telemetry.