DroHubHandler: Set the TelemetryContainer capacity to 1024 and clear it on pop.
As is known the TelemetryContainer is a blocking FIFO
queue which means that when there is no more space it
will block. We do not want that as we are dealing with
real time info. So we increase its size to values where
it is unlikely to get full and flush it whenever the latest
telemetry is popped. This because we have no way to retrieve
the older stale ones, nor we want to