ThriftMessageHandler: Improved ThriftClient.
We no longer have a IOEchoStream data structure and just
pass around the EchoStreams when needed. This so we
increase code locality and also to guarantee that after
a flush we create a new Output EchoStream. It may not be
very efficient but ensures correctness which at the point
of the project is more important.