DroHubHandler: Only initialize video when in INVALID_CONDITION.
This is so that we do not ask multiple times for the permissions,
and thus trigger strange states where multiple video initalizations
are done. So from now on if we are not in INVALID_CONDITION and
we are asked to send live video we just send the current video state
and do not try to initalize any video.