Refresh credentials and freeze config for relay and NAT.
In these changes we finally set all the STUN and TURN servers to
work. The STUN server is required to remove the need to hard code
the NAT address for translation. A wrong address on the nat_1_1_mapping
will lead to wrong results.

Be aware that the ports for the stun and turn servers need to open in
the public facing server.

WE also consider that the wg1 is the interface that connects to the public
facing server