DroHubHandler.java: Fixed field setting not being honored.
There is a quirk in the way that the thrift compiler
generates the bindings in that the fields of a given
class can be set but they are not serialized because the
is_set field is not accordingly set. Instead to make sure
the fields are serialized we need to use a <Field>IsSet()
method. Very unintuitive and this quirk was leading to
data not actually being sent completely.