dashboard.js: Removed telemetry-single-data class.
Removed this class as it was just a pointer for the fact that
the element had a renderer. It turns out we do not need the
class to select telemetry stuff. Instead we just can select for
data-renderer and data-telemetry attributes. This makes it so
we do not need to remember to add the class.