dashboard.js: Do not throw on unreachable range.
We do not fully control the range of values that we
need to render and for some reason they may sometimes
be outside the range we have accounted for. If this is
the case just emit a warning and let the rest of the
page render. Aborting the page rendering due to an
unexpected data is not really acceptable.