DroneMicroServiceManager: Do not cancel database updates.
Even if the connection is gone if we are writing
something to the database we want to finish it. The
previous situation led to the database operations being
canceled and, for example on startup like the IntegrationTests
work the records would never be written leading to incomprehensible
situations where the tests would fail once and always success after
the first failure.