HttpClientHelper: createDevice should honor create_user
Before we had the funny situation where we assumed the
create_user argument of createDevice was false but was ignored
and actually we were always trying to create the user. This did
not fail because at this time it seems our logic to detect
whether the user creation was successful is not working.