Identity Set restrictions on the password.
We now require the minimum size of the password to be
6 characters and have at least 2 unique characters. Also
fixed the tests for the user creation where the user creation
was not correctly tested and reported success on user creation
when it was not the case. The verification of the new requirements
is now validated with the tests.