Identity: Users can have different claims.
Now users have claims assigned to them that allow them
to pass policy checks. In this commit the ActingAdmin,
ActingSubscriber, ActingOwner, ActingPilot and ActingGuest
are possible claims.

The ActingAdmin is the super user ofthe website and is not
bound by business rules.

The ActingSubscriber can check the billing and delegate power
users.

The ActingOwner is the power user and the last tier that can
add users.

The ActingPilot can see the live video as well as control devices.

Finally the ActingGuest can only see live videos or the media section.