User management and administration improvements.
Added the concept of subscription to each user account. The
subscription reflects how the user is monetized and it's limitations in
the system.

Replaced the usage of policy based authorization in favor of claim based
authorization. This led to some deep changes in the user creation and
management code. The reason is that there was too much duplicate
declaration of claims and policies. Also we are able to have one off
claim authorization requirements.