Added support for claim migration.
As the claims available in the system are fluid there needs
to be a way to update current users with the relevant claims
to their basic acting role type. To this end we needed to add
a new field in the DroHubUser called BaseActingType. This value
is then used to synchronize the expected claims of a user of that
type. As the available claims are only changed when there is a new
version of the server, we make this migration at the beginning of
the server startup along with other initializations like database
migration.

DatabaseMigrationExtension.cs was renamed to DroHubMigrationExtension.cs
because we can re-use this cs file to add other migrations, in this
case the claim migration.