feat(Device): Removed UserDevices
Removed the mapping UserDevices table because we no longer want to have
a device associated to a user, but associated to a subscription. This
allows for all the users, besides the guest, to contribute to the
organization device pool. For example if a pilot adds a device then all
his organization will be able to see it, even the guest. Speaking of the
guest, now that the user does not need to take care of each device by
himself, he does not need to create a device, and thus does not require
permissions for it. I am very satisfied with the conceptual direction,
even if this code is yet not tested (When we consolidate we will)

Also refactored some device code to static methods in DroHubUser.cs