fix(InitializeAdminUserExtension): Migration now also migrates subscription for admin
The migration did not take into account the possibility of different
admin configurations needing to be migrated at start up. When there were
different subscriptions or new fields in the admin user nothing was
updated leading to an unhealthy Administrator state. This is important
because if something breaks the administrator needs to be able to fix
small issues.