test: Adapt testing to user roles and subscription
All the user creation needed to be changed to reflect the new concepts
of acting roles, organization and subscription. Test functionality is
the same except that the TestThriftDrone cannot be used with multiple
concurrent mocks as there seems to be races in the Thrift code, which i
cannot be bothered to fix now.
The model used in the Java code for
android should probably be adapted for C#.
A lot of new tests should now
be tried, as the semantics of the roles is not tested, only the previous
cases.