fix(ThriftMessageHandler): Do not propagate Exceptions from ReceiveFromWebSocket
There can be exceptions for things as trivial as a connection being
closed, but we do not do anything with it, so lets just catch the
exception and silence the output slightly.
The exceptions do not matter
because our mechanism for handling connection evens is to issue a
cancellation trigger.