feat(Test, DroHub,App): Application use Token Authentications
Previously we had a system where the application stored and sent the
user and password of drohub when communicating with the backend. That
concept had issues in the way that we needed to signin everytime with
our real crendentials and the credentials needed to be stored
permanently in the application, in a not very secure way. This required
duplicate logic in the App API controller as well as in the
ThriftMessageHander. Now all the token generation and authentication is
done in the UserController.

This changes contain tests for the case
where the guest tries to be a pilot, as well as for the functioning of
the authorization for other acting roles.

There should not be any
visible changes in the frontend, neither of DroHub nor the App.