fix(DevicesController): Action methods have required parameters
Previously there was a lot of room for error with Action arguments that
were options, when in fact they were required. Stop the ambiguity and
have the model validators do the work we were sometimes doing inside the
actions' definitions.

Also non action are now decorated with the
NoneAction attribute.