test(TestCreateAndDeleteDevice): Added support for device creation by non admin user
Previously the devices were always created by the administrator, which
led to logic of ActingType not being able to be tested. At this point it
is still not tested, but the test supports user creation and respective
cleanup, unlike before.

Also relaxed the test's arguments, because
there were a lot of superfluous arguments, whose variation did not
impact any logic.