fix(DevicesController): Do not allow for creation of already existing devices
Since we now have device per subscription instead of per UserDevice it
does not make sense that users can add the same device, if it already
exists. So now the device Create logic is completely independent of
other the Edit Actions. Also created a static nonaction so that it can serve as a
create device helper.

Edit logic also simplified

Also relaxed the requirement of the parameters of the post as we do not yet
care about apperture and other camera related properties.