feat(Whole platform): Auto device discovery and creation
The application now checks if the connected device is known to the
backend or not. If it is not it will direct the user to an activity
where he only needs to name the device. The serial number will
automaticly be added in the Device create request. The device create
request is done through the AndroidApplicationController and uses the
token authentication. User claims checks still apply and tests were
added to verify this was still the case.

In the tests, the createDevice
in HttpClientHelper no longer returns am HttpClientHelper. It does not
make sense as the only cookie a that may be required is from the login.
Other methods do not add anything useful. Probably in the future, more
the HttpClientHelper should be refactored to become simpler and without
the need of Dispose method calls every where.