refactor(android app): Refactor capturer options and peerConnectionParameters
Made the peerConnectionParameters classes be more robust and allow for
their set up at the activity point instead of deeply in the
DroHubHandler. DroHubHandler was moved out of the ThriftConnection to
allow for configuration of the handler by the activity.
With the changes
we are now able to choose the capturer at the activity level. Work can
still be done as the code is still very spaghetti.