parrot-app: Record the gl video instead of screen.
Previously we used the ScreenShare capturer to capture
the video of the drone. This was very not ideal because
when the phone received notifications or switched to other
applications the DroHub also captured the video. This had
massive privacy implications. Now we take the renderer that
Parrot provides and pass it to an Android WebRTC capturer. We
feed the opengl frames to the capturer and therefore we are
able to have all the video rendering, processing and encoding
in hardware. As a consequecne we no longer use the DroHubStreamView
but instead use the SurfaceViewRenderer that is the generic video
renderer for WebRTC. As bonus we are now able to capture the drone
video and deliver it through the same interfaces as other capturers
like the Camera capturer or screen capturer, so theoretically the
user can choose to share the drone, camera or screen and pipe the
real time video to DroHub.