fix(ThriftMessageHandler): Verify the user is in the same subscription as device
Up to now there was no validation that the user actually could send
telemetry to a device which was not registred as is. The only validation
was that the device existed at all. We abandoned the doctrine where
physical posession stealing other people's devices. This is because
somebody could fake a device serial and fake physical ownership.

The
TestWebSocketConnection was changed to test exactly this situation,
where a device belonging to another organization is "piloted" by a
different organization.