DroneMicroServiceManager: Drop Thrift connection after innactivity.
We now have a mechanism to abort the connection if there is
no answer to any action for a given amount of time. The given
amount of time is now 30 seconds. No communication is given to
the user due to the way the cancellation token makes the WebSocket
automatically aborted. In effect the TCP connection is fin/rst.

Added test to verify the actual connection kill.