fix(DroHubUserHelper): isTokenValid now check if passed user is valid.
We were assuming the isTokenValid check was passed a user that is valid
but that may not be the case. Return false as if token is invalid if the user
is not valid. The right response should be an exception but this
would be for a feature not a fix.