fix(ThriftMessageHandler):
Simplified websocket cancellation logic, starting by making the the handler
disposable. This is important because the websocket needs to be closed or
aborted after the handler is disposed. Notice that the handler is instantiated
as a scoped object so we definitely have the ability to control the handler's
shutdown.

Also we now just plain close the connection if we are cancelled or the
connection is not open.

We also do not care if the cancellation token is already cancelled and on
the finalizer always call the Cancel on the token. This is because it is
harmless and it is a simplification on a complicated logic.

Tests are all green