test(IntegrationTest): Delete TestCreateAndLogin
The logic to test if the user can be created or not
needs complete rewriting so that users are created with
different acting types and in the context of subscriptions.

At this point this test was useless, as it did little more
than create users with the admin credentials. There are plenty
of tests adding users already, and this test is misleading.