refactor(Docker): Moved docker directories to a top level docker dir
Now all the docker related directories and files are in a toplevel
docker directory. This is so that the project does not get cluttered
with deployment related files and there is a single place for the docker
topic.

The docker-compose stays in the root of the repo for convenience
but should also be
moved to the docker directory once a task in the
msbuild is created to handle operations in docker-compose.