feat(DroHub): Fixed double device creation and improved device create tests
There was a bug where a user could try to create a device which already
existed. This is an error and against the stealing/editing of existing
devices. As the logic for the device create was not in the DeviceHelper,
now it is and both the DeviceController as well as the
AndroidApplicationController have a common code on device creation. They
also have less responsabilities in preparing the new device to be added,
leaving it all to common code methods in the DeviceHelper.

Due to this
consolidation even DroHubUser extension methods were changed to make the
code simpler.

The improvements above required a more robust framework
for the device creation and disposing, so we now have managed Device
create and delete helper. This allowed for example for the removal of
book keeping in DroneDeviceHelper for device creation and deletion.
Options of device and user creation were removed because they always
were true.

Also added a test where we try to add a same
device twice so that we verify our logic is working.

Split the TestCreateAndDeleteDevice in 2, with the model
validation part being set into TestIncompeteCreateDeviceModelFails.