fix(DroHub): Fixed SignalR broadcasting to all users and test that was not noticing it
The worse cased happened where there was a bug and the test that should
have raised the alarm was not working properly. Specifically the
broadcast signal code was broadcasting to all the users of every
subscription. This instead of broadcasting only to the subscription of
the device received. This commit fixes this and move the code to the
DeviceHelper.

There are several reasons the SignalR checker in
TestThriftDrone was not working. First it was running only one
concurrent thrift session at a time, so only one user, and therefore one
subscription was active, which led to the bug not being actually wrong
(broken clock right once a day). The second, assuming we run the test
with multiple concurrent devices and organizations, was that it was not
checking if the received signalR message was actually containing the
serial number we were looking for. Instead it was just checking for a
match in the timestamp. As the timestamps could be the same due to
instantaneous iterations, the tests passed happily. This plus some more
unfortunate races. In the end adding a check for the serial number as
well as the timestamp led to the test detecting our bug.

Also
refactored the test so that the invariants are removed from the test's
arguments.